+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fix EncryptionSetting.from_response_headers
 * Fix FileVersion.size and FileVersion.mod_time_millis type ambiguity
 * Old buckets (from past tests) are cleaned up before running integration tests in a single thread
+* `B2Api.get_file_info` returns a `FileVersion` object in v2
 
 ## [1.9.0] - 2021-06-07